requirements.txt
openai>=1.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
scipy>=1.11.0
datasets>=2.14.0,<3.0.0
